move intercept cmd
