+++ pkg/client/cli/cliutil/cliutil.go
@@ -1,0 +1,5 @@
+package cliutil
+
+import "regexp"
+
+var HostRx = regexp.MustCompile(`^[a-zA-Z0-9](?:[a-zA-Z0-9\-]*[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9\-]*[a-zA-Z0-9])?)*$`)
