+++ pkg/client/cli/cliutil/cloud_messages_test.go
@@ -1,4 +1,4 @@
-package cli
+package cliutil
 
 import (
 	"context"
