+++ pkg/client/cli/cliutil/update_check_test.go
@@ -1,4 +1,4 @@
-package cli
+package cliutil
 
 import (
 	"bytes"
