+++ pkg/client/cli/cmd_preview.go
@@ -1,15 +1,22 @@
 package cli
 
 import (
+	"bufio"
 	"context"
 	"fmt"
+	"io"
+	"reflect"
+	"strconv"
+	"strings"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"google.golang.org/grpc"
 	empty "google.golang.org/protobuf/types/known/emptypb"
 
+	"github.com/telepresenceio/telepresence/rpc/v2/connector"
 	"github.com/telepresenceio/telepresence/rpc/v2/manager"
+	"github.com/telepresenceio/telepresence/v2/pkg/client/cache"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/cli/cliutil"
 	"github.com/telepresenceio/telepresence/v2/pkg/client/scout"
 )
@@ -97,7 +104,7 @@ func previewCommand() *cobra.Command {
 					if err != nil {
 						return err
 					}
-					fmt.Println(DescribeIntercepts([]*manager.InterceptInfo{intercept}, nil, false))
+					fmt.Println(cliutil.DescribeIntercepts([]*manager.InterceptInfo{intercept}, nil, false))
 					return nil
 				})
 			})
@@ -123,7 +130,7 @@ func previewCommand() *cobra.Command {
 					if err != nil {
 						return err
 					}
-					fmt.Println(DescribeIntercepts([]*manager.InterceptInfo{intercept}, nil, false))
+					fmt.Println(cliutil.DescribeIntercepts([]*manager.InterceptInfo{intercept}, nil, false))
 					return nil
 				})
 			})
@@ -133,4 +140,160 @@ func previewCommand() *cobra.Command {
 	cmd.AddCommand(createCmd, removeCmd)
 
 	return cmd
+}
+
+func selectIngress(
+	ctx context.Context,
+	in io.Reader,
+	out io.Writer,
+	connInfo *connector.ConnectInfo,
+	interceptName string,
+	interceptNamespace string,
+	ingressInfos []*manager.IngressInfo,
+) (*manager.IngressInfo, error) {
+	infos, err := cache.LoadIngressesFromUserCache(ctx)
+	if err != nil {
+		return nil, err
+	}
+	key := connInfo.ClusterServer + "/" + connInfo.ClusterContext
+	selectOrConfirm := "Confirm"
+	cachedIngressInfo := infos[key]
+	if cachedIngressInfo == nil {
+		iis := ingressInfos
+		if len(iis) > 0 {
+			cachedIngressInfo = iis[0] // TODO: Better handling when there are several alternatives. Perhaps use SystemA for this?
+		} else {
+			selectOrConfirm = "Select" // Hard to confirm unless there's a default.
+			if interceptNamespace == "" {
+				interceptNamespace = "default"
+			}
+			cachedIngressInfo = &manager.IngressInfo{
+				// Default Settings
+				Host:   fmt.Sprintf("%s.%s", interceptName, interceptNamespace),
+				Port:   80,
+				UseTls: false,
+			}
+		}
+	}
+
+	reader := bufio.NewReader(in)
+
+	fmt.Fprintf(out, "\n"+ingressDesc+"\n", selectOrConfirm)
+	reply := &manager.IngressInfo{}
+	if reply.Host, err = askForHost(ingressQ1, cachedIngressInfo.Host, reader, out); err != nil {
+		return nil, err
+	}
+	if reply.Port, err = askForPortNumber(cachedIngressInfo.Port, reader, out); err != nil {
+		return nil, err
+	}
+	if reply.UseTls, err = askForUseTLS(cachedIngressInfo.UseTls, reader, out); err != nil {
+		return nil, err
+	}
+	if cachedIngressInfo.L5Host == "" {
+		cachedIngressInfo.L5Host = reply.Host
+	}
+	if reply.L5Host, err = askForHost(ingressQ4, cachedIngressInfo.L5Host, reader, out); err != nil {
+		return nil, err
+	}
+	fmt.Fprintln(out)
+
+	if !ingressInfoEqual(cachedIngressInfo, reply) {
+		infos[key] = reply
+		if err = cache.SaveIngressesToUserCache(ctx, infos); err != nil {
+			return nil, err
+		}
+	}
+	return reply, nil
+}
+
+func ingressInfoEqual(a, b *manager.IngressInfo) bool {
+	return a.Host == b.Host && a.L5Host == b.L5Host && a.Port == b.Port && a.UseTls == b.UseTls
+}
+
+const (
+	ingressDesc = `To create a preview URL, telepresence needs to know how requests enter
+	your cluster.  Please %s the ingress to use.`
+	ingressQ1 = `1/4: What's your ingress' IP address?
+	 You may use an IP address or a DNS name (this is usually a
+	 "service.namespace" DNS name).`
+	ingressQ2 = `2/4: What's your ingress' TCP port number?`
+	ingressQ3 = `3/4: Does that TCP port on your ingress use TLS (as opposed to cleartext)?`
+	ingressQ4 = `4/4: If required by your ingress, specify a different hostname
+	 (TLS-SNI, HTTP "Host" header) to be used in requests.`
+)
+
+func showPrompt(out io.Writer, question string, defaultValue any) {
+	if reflect.ValueOf(defaultValue).IsZero() {
+		fmt.Fprintf(out, "\n%s\n\n       [no default]: ", question)
+	} else {
+		fmt.Fprintf(out, "\n%s\n\n       [default: %v]: ", question, defaultValue)
+	}
+}
+
+func askForHost(question, cachedHost string, reader *bufio.Reader, out io.Writer) (string, error) {
+	for {
+		showPrompt(out, question, cachedHost)
+		reply, err := reader.ReadString('\n')
+		if err != nil {
+			return "", err
+		}
+		reply = strings.TrimSpace(reply)
+		if reply == "" {
+			if cachedHost == "" {
+				continue
+			}
+			return cachedHost, nil
+		}
+		if cliutil.HostRx.MatchString(reply) {
+			return reply, nil
+		}
+		fmt.Fprintf(out,
+			"Address %q must match the regex [a-z0-9]([-a-z0-9]*[a-z0-9])?(\\.[a-z0-9]([-a-z0-9]*[a-z0-9])?)* (e.g. 'myingress.mynamespace')\n",
+			reply)
+	}
+}
+
+func askForPortNumber(cachedPort int32, reader *bufio.Reader, out io.Writer) (int32, error) {
+	for {
+		showPrompt(out, ingressQ2, cachedPort)
+		reply, err := reader.ReadString('\n')
+		if err != nil {
+			return 0, err
+		}
+		reply = strings.TrimSpace(reply)
+		if reply == "" {
+			if cachedPort == 0 {
+				continue
+			}
+			return cachedPort, nil
+		}
+		port, err := strconv.Atoi(reply)
+		if err == nil && port > 0 {
+			return int32(port), nil
+		}
+		fmt.Fprintln(out, "port must be a positive integer")
+	}
+}
+
+func askForUseTLS(cachedUseTLS bool, reader *bufio.Reader, out io.Writer) (bool, error) {
+	yn := "n"
+	if cachedUseTLS {
+		yn = "y"
+	}
+	showPrompt(out, ingressQ3, yn)
+	for {
+		reply, err := reader.ReadString('\n')
+		if err != nil {
+			return false, err
+		}
+		switch strings.TrimSpace(reply) {
+		case "":
+			return cachedUseTLS, nil
+		case "n", "N":
+			return false, nil
+		case "y", "Y":
+			return true, nil
+		}
+		fmt.Fprintf(out, "       please answer 'y' or 'n'\n       [default: %v]: ", yn)
+	}
 }
