+++ pkg/client/userd/grpc.go
@@ -468,7 +468,7 @@ func (s *Service) Quit(ctx context.Context, _ *empty.Empty) (*empty.Empty, error
 
 func (s *Service) ListCommands(ctx context.Context, _ *empty.Empty) (groups *rpc.CommandGroups, err error) {
 	s.logCall(ctx, "ListCommands", func(ctx context.Context) {
-		groups, err = cliutil.CommandsToRPC(s.getCommands()), nil
+		groups, err = cliutil.CommandsToRPC(s.getCommands(ctx)), nil
 	})
 	return
 }
@@ -478,7 +478,7 @@ func (s *Service) RunCommand(ctx context.Context, req *rpc.RunCommandRequest) (r
 		cmd := &cobra.Command{
 			Use: "fauxmand",
 		}
-		cli.AddCommandGroups(cmd, s.getCommands())
+		cli.AddCommandGroups(cmd, s.getCommands(ctx))
 		args := req.GetOsArgs()
 		cmd.SetArgs(args)
 		cmd, args, err = cmd.Find(args)
@@ -490,18 +490,25 @@ func (s *Service) RunCommand(ctx context.Context, req *rpc.RunCommandRequest) (r
 		cmd.SetOut(outW)
 		cmd.SetErr(errW)
 
+		if _, ok := cmd.Annotations[commands.CommandRequiresConnectorServer]; ok {
+			ctx = commands.WithConnectorServer(ctx, s)
+		}
+
 		if _, ok := cmd.Annotations[commands.CommandRequiresSession]; ok {
 			err = s.withSession(ctx, "cmd-"+cmd.Name(), func(ctx context.Context, s trafficmgr.Session) error {
 				ctx = commands.WithCwd(ctx, req.GetCwd())
+				ctx = commands.WithSession(ctx, s)
 				return cmd.ExecuteContext(ctx)
 			})
 		} else {
 			ctx = commands.WithCwd(ctx, req.GetCwd())
 			err = cmd.ExecuteContext(ctx)
 		}
-		if err != nil {
+
+		if err = cmd.ExecuteContext(ctx); err != nil {
 			return
 		}
+
 		result = &rpc.RunCommandResponse{
 			Stdout: outW.Bytes(),
 			Stderr: errW.Bytes(),
